+++ node2vec_spark/src/test/scala/au/csiro/data61/randomwalk/efficient/GraphMapTest.scala
@@ -6,11 +6,11 @@ import org.scalatest.FunSuite
 class GraphMapTest extends FunSuite {
 
   test("test GraphMap data structure") {
-    val e1 = Array(Edge(1, 2, 1.0))
-    val e2 = Array(Edge(1, 3, 1.0))
-    val e3 = Array(Edge(2, 3, 1.0))
-    val e4 = Array(Edge(2, 1, 1.0))
-    var v2N = Array((1L, e1))
+    val e1 = Array(Edge(1, 2, 1.0f))
+    val e2 = Array(Edge(1, 3, 1.0f))
+    val e3 = Array(Edge(2, 3, 1.0f))
+    val e4 = Array(Edge(2, 1, 1.0f))
+    var v2N = Array((1, e1))
     GraphMap.addVertex(1, e1)
     GraphMap.addVertex(2)
     assert(GraphMap.getNumEdges == 1)
@@ -18,7 +18,7 @@ class GraphMapTest extends FunSuite {
     assertMap(v2N, GraphMap)
 
     GraphMap.reset
-    v2N = Array((1L, e1 ++ e2))
+    v2N = Array((1, e1 ++ e2))
     GraphMap.addVertex(1, e1 ++ e2)
     GraphMap.addVertex(2)
     GraphMap.addVertex(3)
@@ -26,19 +26,19 @@ class GraphMapTest extends FunSuite {
 
 
     GraphMap.reset
-    v2N = v2N ++ Array((2L, e3 ++ e4))
+    v2N = v2N ++ Array((2, e3 ++ e4))
     GraphMap.addVertex(2, e3 ++ e4)
     GraphMap.addVertex(1, e1 ++ e2)
     GraphMap.addVertex(3)
     assertMap(v2N, GraphMap)
   }
 
-  private def assertMap(verticesToNeighbors: Array[(Long, Array[Edge[Double]])], gMap: GraphMap
+  private def assertMap(verticesToNeighbors: Array[(Int, Array[Edge[Float]])], gMap: GraphMap
     .type) = {
     for (v <- verticesToNeighbors) {
-      var neighbors: Array[(Long, Double)] = Array()
+      var neighbors: Array[(Int, Float)] = Array()
       for (e <- v._2) {
-        neighbors = neighbors ++ Array((e.dstId, e.attr))
+        neighbors = neighbors ++ Array((e.dstId.toInt, e.attr))
       }
       assert(gMap.getNeighbors(v._1) sameElements neighbors)
     }
